+++ libsolidity/analysis/SyntaxChecker.h
@@ -52,10 +52,10 @@ private:
 	virtual bool visit(ModifierDefinition const& _modifier) override;
 	virtual void endVisit(ModifierDefinition const& _modifier) override;
 
-	// Reports an error if _statement is a VariableDeclarationStatement.
-	// Used by if/while/for to check for single statement variable declarations
-	// without a block.
-	void checkSingleStatementVariableDeclaration(ASTNode const* _statement);
+	/// Reports an error if _statement is a VariableDeclarationStatement.
+	/// Used by if/while/for to check for single statement variable declarations
+	/// without a block.
+	void checkSingleStatementVariableDeclaration(ASTNode const& _statement);
 
 	virtual bool visit(IfStatement const& _ifStatement) override;
 	virtual bool visit(WhileStatement const& _whileStatement) override;